Moved the filesystem commands into their own package, unsync_fs
