+++ unsync_fs/unsync_fs/__init__.py
@@ -1,0 +1,6 @@
+unsync_prefix = 'filesystem'
+
+unsync_commands = [
+    'files.list_files',
+    'files.shcopy_files',
+]
